requirements.txt
numpy
Pillow
matplotlib
pytest
